+++ relay-server/src/actors/events.rs
@@ -673,6 +673,7 @@ impl EventProcessor {
 
             // session data is never considered as part of deduplication
             ItemType::Session => false,
+            ItemType::SessionAggregates => false,
         }
     }