+++ relay-server/src/endpoints/common.rs
@@ -348,6 +348,8 @@ fn check_envelope_size_limits(config: &Config, envelope: &Envelope) -> bool {
                 attachments_size += item.len()
             }
             ItemType::Session => session_count += 1,
+            // TODO: extract the actual session counts from the aggregates
+            ItemType::SessionAggregates => session_count += 1,
             ItemType::UserReport => (),
         }
     }