+++ relay-server/src/envelope.rs
@@ -91,6 +91,8 @@ pub enum ItemType {
     UserReport,
     /// Session update data.
     Session,
+    /// Aggregated session data.
+    SessionAggregates,
 }
 
 impl ItemType {
@@ -118,6 +120,7 @@ impl fmt::Display for ItemType {
             Self::UnrealReport => write!(f, "unreal report"),
             Self::UserReport => write!(f, "user feedback"),
             Self::Session => write!(f, "session"),
+            Self::SessionAggregates => write!(f, "session aggregates"),
         }
     }
 }
@@ -500,7 +503,7 @@ impl Item {
             ItemType::FormData => false,
 
             // The remaining item types cannot carry event payloads.
-            ItemType::UserReport | ItemType::Session => false,
+            ItemType::UserReport | ItemType::Session | ItemType::SessionAggregates => false,
         }
     }
 
@@ -518,6 +521,7 @@ impl Item {
             ItemType::UnrealReport => true,
             ItemType::UserReport => true,
             ItemType::Session => false,
+            ItemType::SessionAggregates => false,
         }
     }
 }